Switch storybook framework to react vite
