+++ packages/keychain/package.json
@@ -40,7 +40,6 @@
     "graphql": "^16.3.0",
     "graphql-request": "^5.0.0",
     "js-cookie": "^3.0.1",
-    "next": "^14.2.5",
     "p-throttle": "^6.2.0",
     "posthog-js": "^1.181.0",
     "react": "^18.3.1",
@@ -60,8 +59,8 @@
     "@storybook/addon-interactions": "^8.2.9",
     "@storybook/addon-links": "^8.2.9",
     "@storybook/blocks": "^8.2.9",
-    "@storybook/nextjs": "^8.2.9",
     "@storybook/react": "^8.2.9",
+    "@storybook/react-vite": "^8.4.7",
     "@storybook/test": "^8.2.5",
     "@storybook/test-runner": "^0.19.1",
     "@testing-library/react": "^13.4.0",
@@ -72,7 +71,6 @@
     "autoprefixer": "^10.4.18",
     "concurrently": "^9.0.1",
     "eslint": "^9.12.0",
-    "eslint-config-next": "^12.2.5",
     "eslint-plugin-storybook": "^0.6.13",
     "http-server": "^14.1.1",
     "jsdom": "^25.0.1",
