+++ packages/keychain/src/components/layout/Footer/index.tsx
@@ -2,7 +2,7 @@ import { HStack, VStack, Text } from "@chakra-ui/react";
 import { CartridgeLogo } from "@cartridge/ui";
 import React, { useEffect, useRef } from "react";
 import { FOOTER_HEIGHT, useLayout } from "@/components/layout";
-import NextLink from "next/link";
+import { Link } from "react-router-dom";
 
 export function Footer({
   children,
@@ -66,8 +66,8 @@ export function Footer({
           h={FOOTER_HEIGHT / 4}
           gap={1}
           opacity={0.5}
-          as={NextLink}
-          href="https://cartridge.gg"
+          as={Link}
+          to="https://cartridge.gg"
           target="_blank"
           overflow="hidden"
           _hover={{
