+++ packages/keychain/src/pages/slot/consent.tsx
@@ -3,9 +3,8 @@ import { Button, Text } from "@chakra-ui/react";
 import { Container, Footer } from "@/components/layout";
 import { useRouter } from "next/router";
 import { useCallback, useEffect } from "react";
-import dynamic from "next/dynamic";
 
-function Consent() {
+export function Consent() {
   const router = useRouter();
 
   const onSubmit = useCallback(async () => {
@@ -54,5 +53,3 @@ function Consent() {
     </Container>
   );
 }
-
-export default dynamic(() => Promise.resolve(Consent), { ssr: false });
